+++ src/pymmcore_widgets/control/_stage_explorer/_stage_explorer.py
@@ -5,6 +5,7 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Callable, cast
 
 import numpy as np
+import useq
 import vispy.scene
 from pymmcore_plus import CMMCorePlus, Keyword
 from qtpy.QtCore import QPoint, Qt
@@ -21,12 +22,12 @@ from qtpy.QtWidgets import (
 from superqt import QIconifyIcon
 
 from pymmcore_widgets.control._q_stage_controller import QStageMoveAccumulator
+from pymmcore_widgets.control._rois.roi_manager import SceneROIManager
 
 from ._stage_position_marker import StagePositionMarker
 from ._stage_viewer import StageViewer, get_vispy_scene_bounds
 
 if TYPE_CHECKING:
-    import useq
     from PyQt6.QtGui import QAction, QActionGroup
     from qtpy.QtCore import QTimerEvent
     from vispy.app.canvas import MouseEvent
@@ -46,6 +47,9 @@ CLEAR = "Clear View"
 SNAP = "Snap on Double Click"
 POLL_STAGE = "Show FOV Position"
 SHOW_GRID = "Show Grid"
+ROIS = "Create ROI (option click)"
+DELETE_ROIS = "Delete All ROIs"
+SCAN = "Scan Selected ROIs"
 
 
 # this might belong in _stage_position_marker.py
@@ -142,6 +146,7 @@ class StageExplorer(QWidget):
 
         self._stage_viewer = StageViewer(self)
         self._stage_viewer.setCursor(Qt.CursorShape.CrossCursor)
+        self.roi_manager = SceneROIManager(self._stage_viewer.canvas)
 
         self._grid_lines = vispy.scene.GridLines(
             parent=self._stage_viewer.view.scene,
@@ -155,7 +160,7 @@ class StageExplorer(QWidget):
 
         # properties
         self._auto_zoom_to_fit: bool = False
-        self._snap_on_double_click: bool = False
+        self._snap_on_double_click: bool = True
         self._poll_stage_position: bool = True
 
         # stage position marker mode
@@ -184,6 +189,9 @@ class StageExplorer(QWidget):
             SNAP: ("mdi:camera-outline", True, self._on_snap_action),
             POLL_STAGE: ("mdi:map-marker-outline", True, self._on_poll_stage_action),
             SHOW_GRID: ("mdi:grid", True, self._on_show_grid_action),
+            # ROIS: ("mdi:vector-square", True, None),
+            DELETE_ROIS: ("mdi:vector-square-remove", False, self._remove_rois),
+            SCAN: ("iconoir:path-arrow-solid", False, self._on_scan_action),
         }
         # fmt: on
 
@@ -193,7 +201,8 @@ class StageExplorer(QWidget):
                 icon = QIconifyIcon(icon, color=GRAY)
             self._actions[a_text] = action = QAction(icon, a_text, self)
             action.setCheckable(check)
-            action.triggered.connect(callback)
+            if callback is not None:
+                action.triggered.connect(callback)
 
             if a_text == POLL_STAGE:
                 # create special toolbutton with a context menu on right-click
@@ -201,9 +210,13 @@ class StageExplorer(QWidget):
                 btn.setDefaultAction(action)
                 self._toolbar.addWidget(btn)
                 action.setChecked(self._poll_stage_position)
-                self._on_poll_stage_action(self._poll_stage_position)
             else:
                 self._toolbar.addAction(action)
+        self._toolbar.addActions(self.roi_manager.mode_actions.actions())
+        # update checked state of the actions
+        self._on_poll_stage_action(self._poll_stage_position)
+        self._on_snap_action(self._snap_on_double_click)
+        self._actions[SCAN].setEnabled(False)
 
         # add stage pos label to the toolbar
         self._stage_pos_label = QLabel()
@@ -351,7 +364,32 @@ class StageExplorer(QWidget):
                 self._position_indicator.show_marker
             )
 
+    def _remove_rois(self) -> None:
+        """Delete all the ROIs."""
+        self.roi_manager.clear()
+
+    def _on_scan_action(self) -> None:
+        """Scan the selected ROIs."""
+        if not (active_rois := self.roi_manager.selected_rois()):
+            return
+        active_roi = active_rois[0]
+        fov_w, fov_h, z_pos = self._fov_w_h_z_pos()
+        pos = active_roi.create_useq_position(fov_w=fov_w, fov_h=fov_h, z_pos=z_pos)
+        seq = useq.MDASequence(stage_positions=[pos])
+
+        if not self._mmc.mda.is_running():
+            self._our_mda_running = True
+            self._mmc.run_mda(seq)
+
     # CORE ------------------------------------------------------------------------
+
+    def _fov_w_h_z_pos(self) -> tuple[float, float, float]:
+        """Return the field of view width, height and z position."""
+        px = self._mmc.getPixelSizeUm()
+        fov_w = self._mmc.getImageWidth() * px
+        fov_h = self._mmc.getImageHeight() * px
+        z_pos = self._mmc.getZPosition()
+        return fov_w, fov_h, z_pos
 
     def _on_sys_config_loaded(self) -> None:
         """Clear the scene when the system configuration is loaded."""
@@ -364,6 +402,8 @@ class StageExplorer(QWidget):
     def _on_mouse_double_click(self, event: MouseEvent) -> None:
         """Move the stage to the clicked position."""
         if not self._mmc.getXYStageDevice():
+            return
+        if self.roi_manager.mode == "create-poly":
             return
 
         # map the clicked canvas position to the stage position
