+++ src/pymmcore_widgets/control/_stage_explorer/_stage_viewer.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, cast
-from unittest.mock import patch
+from typing import TYPE_CHECKING, cast
 
 import cmap
 import numpy as np
@@ -23,25 +22,6 @@ if TYPE_CHECKING:
     class VisualNode(vispy.scene.Node, vispy.visuals.Visual): ...
 
 
-class KeylessSceneCanvas(vispy.scene.SceneCanvas):
-    """Steal all key events from vispy."""
-
-    def create_native(self) -> None:
-        from vispy.app.backends._qt import CanvasBackendDesktop
-
-        class CustomCanvasBackend(CanvasBackendDesktop):
-            def keyPressEvent(self, ev: Any) -> None:
-                QWidget.keyPressEvent(self, ev)
-
-            def keyReleaseEvent(self, ev: Any) -> None:
-                QWidget.keyPressEvent(self, ev)
-
-        with patch.object(
-            self._app.backend_module, "CanvasBackend", CustomCanvasBackend
-        ):
-            super().create_native()
-
-
 class StageViewer(QWidget):
     """A widget to add images with a transform to a vispy canves."""
 
@@ -53,7 +33,7 @@ class StageViewer(QWidget):
         self._clims: tuple[float, float] | None = None
         self._cmap: cmap.Colormap = cmap.Colormap("gray")
 
-        self.canvas = KeylessSceneCanvas(show=True)
+        self.canvas = vispy.scene.SceneCanvas(show=True)
 
         self.view = cast("ViewBox", self.canvas.central_widget.add_view())
         self.view.camera = scene.PanZoomCamera(aspect=1)
